// client/assets/scripts/logic/red/RedDotType.ts
export enum RedDotType {
    empty = 0,
    testLogin = 1,  //测试
    HeroupLevel = 2,//升级
    HeroupStar = 3,//升星
    HeroResonanceLevel = 4,  //共鸣等级
    HeroresonanceStar = 5,  //共鸣星级
    HeroRecommend = 6,  //推荐阵容
    HeroAutoAscend = 7, //一键升星
    HeroResolve = 8,//一键遣散
    HeroPainting = 9,//绘卷
    StageFirstReward = 10,//关卡首通奖励
    HeroDrug = 11,//捕虫红点
    HeroGene = 12,//基因红点
    HeroGacha = 13,//保底抽红点
    HeroBook = 14,//图签红点
    GachaAds = 15,//抽卡广告红点
    Welfare = 16,//福利活动
    RedStamina = 17,//体力红点
    NewPlayerSignIn = 18, //新服签到
    SignIn = 19, //七日签到
    HeroRoad = 20, //舰队启航
    FriendApply = 30, //好友申请
    FriendValueReceive = 31, //友情点领取
    MonthlyCard = 32, //月卡
    Vip = 33,       //vip
    GoldBuy = 34,//购买金币
    Prestige = 35, //冒险日志
    Mail = 36, //邮件
    DayTask = 37, //每日任务
    WeekTask = 38, //每周任务
    AchievementTask = 39, //成就任务
    ChallengeDailyFreeTimes = 49,//挑战-日常探险-免费次数
    ChallengeDailyAward = 50,//挑战-日常探险-有可领取的奖励
    ChapterAward = 51,//关卡奖励红点
    ClimbingTowerChallenge = 52,//精英挑战次数红点
    ClimbingTowerDailyReward = 53, //精英挑战每天奖励
    ClimbingTowerStageReward = 54, //精英挑战阶段奖励
    EveryDayChallengeFreeNum = 55, //每日挑战次数
    EveryDayChallengeUpReward = 56, //每日挑战提示奖励
    EveryDayChallengeBoxReward = 57, //每日box奖励
    TopWarChallengeFreeNum = 58,   //生化危机免费挑战次数
    Weapon_Job_Archer = 59,   //武器图鉴射手
    Weapon_Job_Assassin = 60, //武器图鉴刺客 
    Weapon_Job_Priest = 61,   //武器图鉴牧师
    Weapon_Job_Caster = 62,   //武器图鉴法师
    Weapon_Job_Warrior = 63,  //武器图鉴战士
    Battle_Pass = 64,       //战令红点
    Bag_Compose = 65,       //背包合成红点
    Hand_book = 66,         //指南红点
    // Book_Job_Archer=67,   //武器射手
    // Book_Job_Assassin=68, //武器刺客 
    // Book_Job_Priest=69,   //武器牧师
    // Book_Job_Caster=70,   //武器法师
    // Book_Job_Warrior=71,  //武器战士

    BooK_Equip = 67,        //武器装备
    Book_collect = 68,     //收藏
    Book_Job = 69,
    First_Recharge = 70, //首充红点
    Seven_Gift_Pack = 71,//七日礼包
    Welfare_Open = 72,  //福利开启
    Wear_Jade = 73,     //可穿戴羽毛
    Wear_Equip = 74,    //可一键穿戴装备
    Feng_Yun_Rank = 75,  //风云榜红点
    notice = 76,            //公告红点
    Head_Icon_Red = 77,
    Weapon_recovery = 78,  //武器回收
    Vip_Buy = 79, //VIP购买
    Guild_Skill = 80,//帮会技能红点
    Guild_Task = 81, //帮会任务
    Guild_Apply = 82,//有人申请公会红点
    Guild_Sign = 83,//公会是否可以签到
    Guild_Activity = 84,//公会可以砍一刀
    TenGacha = 85,//十连红点
    SignGiftRed = 86,//百抽红点
    Combine_Pass = 87,//组合战令
    Combine_Grow = 88,//组合养成
    Combine_Shop = 89,//组合商店
    Gacha_Up_red = 90,//抽卡up红点
    HeroReplace = 91,//有可替换的英雄
    TenBookGacha = 92,//十连武器抽
    HeroBagRed = 93,//背包红点
    Equip_Strengthen = 94,//装备可以一键强化
    Free_Fight_Token = 95,//PVP有免费挑战次数时（功能开启后）
    PVP_Fight_Team = 96,//有空的英雄 有空的武器
    Combine_Recharge = 97,//累充活动
}